move logger to library
